+++ packages/blade-svelte/src/components/Button/IconButton/BaseIconButton/index.ts
@@ -1,0 +1,2 @@
+export { default as BaseIconButton } from './BaseIconButton.svelte';
+export type { BaseIconButtonProps } from './types';