+++ packages/blade-svelte/src/components/Icons/index.ts
@@ -3,19 +3,29 @@ export * from './types';
 
 // Stroked Icons
 export { AlertOctagonIcon } from './AlertOctagonIcon';
+export { ArrowLeftIcon } from './ArrowLeftIcon';
 export { AlertTriangleIcon } from './AlertTriangleIcon';
 export { CheckCircleIcon } from './CheckCircleIcon';
 export { CheckIcon } from './CheckIcon';
 export { ChevronDownIcon } from './ChevronDownIcon';
+export { ChevronLeftIcon } from './ChevronLeftIcon';
+export { ChevronRightIcon } from './ChevronRightIcon';
+export { ChevronUpDownIcon } from './ChevronUpDownIcon';
 export { HomeIcon } from './HomeIcon';
 export { CloseIcon } from './CloseIcon';
 export { CreditCardIcon } from './CreditCardIcon';
 export { InfoIcon } from './InfoIcon';
+export { PhoneIcon } from './PhoneIcon';
 export { PlusIcon } from './PlusIcon';
 export { SearchIcon } from './SearchIcon';
 
 export { UserIcon } from './UserIcon';
 export { BuildingIcon } from './BuildingIcon';
+export { BankIcon } from './BankIcon';
+export { LockIcon } from './LockIcon';
+
+// Branded Icons
+export { RazorpayTrustIcon } from './RazorpayTrustIcon';
 
 // Filled Icons
 export { MoreFilledIcon } from './MoreFilledIcon';